chore: restructure and focus only on usgs
