+++ lerna.json
@@ -1,6 +1,5 @@
 {
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.0.0"
+  "packages": ["packages/*"],
+  "version": "0.0.0",
+  "npmClient": "yarn"
 }
