--- packages/dwr/package.json
@@ -1,51 +1,0 @@
-{
-  "version": "0.1.0",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "engines": {
-    "node": ">=10"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test",
-    "test:watch": "tsdx test --watchAll",
-    "lint": "tsdx lint",
-    "prepare": "tsdx build"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "tsdx lint"
-    }
-  },
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5"
-  },
-  "name": "streamster-dwr",
-  "author": "Ben Tyler",
-  "repository": "git://github.com/tylerben/streamster.git",
-  "module": "dist/streamster.esm.js",
-  "devDependencies": {
-    "@types/papaparse": "^5.0.3",
-    "husky": "^4.2.5",
-    "lerna": "^3.22.1",
-    "tsdx": "^0.13.2",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.5"
-  },
-  "dependencies": {
-    "axios": "^0.19.2",
-    "papaparse": "^5.2.0"
-  }
-}
