--- packages/dwr/src/dwr/SurfaceWater/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SurfaceWater';
