--- packages/dwr/src/dwr/types.ts
@@ -1,5 +1,0 @@
-import { dwrSurfaceWaterService } from './SurfaceWater/types';
-
-export interface dwrService {
-  surfaceWater(): dwrSurfaceWaterService;
-}
