+++ packages/usgs/src/Daily/Daily.test.ts
@@ -1,10 +1,10 @@
-import Streamster from '../../index';
+import usgs from '../index';
 
 describe('USGS Daily Service Tests: Time Series', () => {
   test('Single Site (Specific Dates): Returns data', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           sites: '09361500',
@@ -23,7 +23,7 @@ describe('USGS Daily Service Tests: Time Series', () => {
   test('Single State (Specific Dates): Returns data', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           stateCd: 'co',
@@ -42,7 +42,7 @@ describe('USGS Daily Service Tests: Time Series', () => {
   test('Single HUC (Specific Dates): Returns data in expected shape', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           huc: '14080104',
@@ -61,7 +61,7 @@ describe('USGS Daily Service Tests: Time Series', () => {
   test('Bounding Box (Specific Dates): Returns data in expected shape', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           bBox: '-108.115341,37.031969,-107.385747,37.424825',
@@ -80,7 +80,7 @@ describe('USGS Daily Service Tests: Time Series', () => {
   test('Single County (Specific Dates): Returns data in expected shape', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           countyCd: '08067',
@@ -99,7 +99,7 @@ describe('USGS Daily Service Tests: Time Series', () => {
   test('Single Site (Period): Returns data in expected shape', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.daily().getDailyData({
         format: 'raw',
         queryParameters: {
           sites: '09361500',
