+++ packages/usgs/src/Daily/Daily.ts
