+++ packages/usgs/src/Daily/index.ts
