+++ packages/usgs/src/Instantaneous/Instantaneous.test.ts
@@ -1,17 +1,17 @@
-import Streamster from '../../index';
+import usgs from '../index';
 
-describe('USGS Daily Service Tests: Time Series', () => {
+describe('USGS instantaneous Service Tests: Time Series', () => {
   test('Single Site (Specific Dates): Returns data', async () => {
     jest.setTimeout(10000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           sites: '09361500',
           siteStatus: 'active',
           parameterCd: '00060',
-          startDT: '2019-10-01',
-          endDT: '2019-10-02',
+          startDT: '2019-10-01T00:00-0700',
+          endDT: '2019-10-02T23:59-0700',
         },
       });
       expect(data.value.timeSeries.length).toBeGreaterThan(0);
@@ -21,16 +21,16 @@ describe('USGS Daily Service Tests: Time Series', () => {
   });
 
   test('Single State (Specific Dates): Returns data', async () => {
-    jest.setTimeout(10000);
+    jest.setTimeout(30000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           stateCd: 'co',
           siteStatus: 'active',
           parameterCd: '00060',
-          startDT: '2019-10-01',
-          endDT: '2019-10-02',
+          startDT: '2019-10-01T00:00-0700',
+          endDT: '2019-10-02T23:59-0700',
         },
       });
       expect(data.value.timeSeries.length).toBeGreaterThan(0);
@@ -40,16 +40,16 @@ describe('USGS Daily Service Tests: Time Series', () => {
   });
 
   test('Single HUC (Specific Dates): Returns data in expected shape', async () => {
-    jest.setTimeout(10000);
+    jest.setTimeout(30000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           huc: '14080104',
           siteStatus: 'active',
           parameterCd: '00060',
-          startDT: '2019-10-01',
-          endDT: '2019-10-01',
+          startDT: '2019-10-01T00:00-0700',
+          endDT: '2019-10-02T23:59-0700',
         },
       });
       expect(data.value.timeSeries.length).toBeGreaterThan(0);
@@ -59,16 +59,16 @@ describe('USGS Daily Service Tests: Time Series', () => {
   });
 
   test('Bounding Box (Specific Dates): Returns data in expected shape', async () => {
-    jest.setTimeout(10000);
+    jest.setTimeout(30000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           bBox: '-108.115341,37.031969,-107.385747,37.424825',
           siteStatus: 'active',
           parameterCd: '00060',
-          startDT: '2019-10-01',
-          endDT: '2019-10-01',
+          startDT: '2019-10-01T00:00-0700',
+          endDT: '2019-10-02T23:59-0700',
         },
       });
       expect(data.value.timeSeries.length).toBeGreaterThan(0);
@@ -78,16 +78,16 @@ describe('USGS Daily Service Tests: Time Series', () => {
   });
 
   test('Single County (Specific Dates): Returns data in expected shape', async () => {
-    jest.setTimeout(10000);
+    jest.setTimeout(30000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           countyCd: '08067',
           siteStatus: 'active',
           parameterCd: '00060',
-          startDT: '2019-10-01',
-          endDT: '2019-10-01',
+          startDT: '2019-10-01T00:00-0700',
+          endDT: '2019-10-02T23:59-0700',
         },
       });
       expect(data.value.timeSeries.length).toBeGreaterThan(0);
@@ -97,9 +97,9 @@ describe('USGS Daily Service Tests: Time Series', () => {
   });
 
   test('Single Site (Period): Returns data in expected shape', async () => {
-    jest.setTimeout(10000);
+    jest.setTimeout(30000);
     try {
-      const data = await Streamster.usgs.daily().getDailyData({
+      const data = await usgs.instantaneous().getInstantaneousData({
         format: 'raw',
         queryParameters: {
           sites: '09361500',
