+++ packages/usgs/src/Instantaneous/Instantaneous.ts
