--- packages/usgs/src/dwr/index.ts
@@ -1,11 +1,0 @@
-import { dwrService } from './types';
-import { dwrSurfaceWaterService } from './SurfaceWater/types';
-import SurfaceWater from './SurfaceWater';
-
-const dwr: dwrService = {
-  surfaceWater(): dwrSurfaceWaterService {
-    return new SurfaceWater();
-  },
-};
-
-export default dwr;
