+++ packages/usgs/src/static/parameters.json
