+++ packages/usgs/src/static/sites.json
