+++ packages/usgs/src/types.ts
