--- packages/usgs/src/usgs/Daily/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Daily';
