--- packages/usgs/src/usgs/Instantaneous/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Instantaneous';
